requirements.txt
Django==5.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.7
python-decouple==3.8
django-cors-headers==4.3.1